osmo-bts-sysmo/oml.c: Update opstate and avstate after rf_lock completes
According to 3GPP TS 12.21 Ch. 5.3:
The admin state can be used to control the power of a trx by setting it
to locked or unlocked.
As a result when the operation completes the trx should set its
operational state to disabled and the availability to off-line (if
locked) or enabled and ok (if unlocked).

Before this patch that didn't happen and so the nitb would still list
all trx as operational/ok even after the trx has been switched off.

The spec doesn't clearly recommend setting opstate to disabled, but that
is what the nanoBTS does, so mimic its behaviour.