Add functionality to interoperate with OsmocomBB virt-PHY.
This patch improves the virtual physical layer designed to simulate the
air interface. The purpose is to get rid of the hardware requirements
and be able to start testing and implementing layer 2 communication
functionality on one machine. Multicast sockets are used to enable
bidirectional communication between the BTS and the MS process.
The GSMTAP protocol designed for wireshark capturing is used to
encapsulate the payload on the virtual physical layer.

* Working multicast socket communication and extraction of its
  functionality.
* Fixed OML and RSL startup sequences.
* Encapsulation and parsing methods to and from GSMTAP messages.
* Basic handlers for file descriptor callbacks from incoming multicast
  messages.
* Multiplexing to different channels based on GSMTAP header channel
  type.

Change-Id: Ieed4930961041c6e9bce5d23c9519ad1e56f676e

fixup

Change-Id: I16ff53a713f523821d05caced17ee75863276def